sqlite3 for Registering and login
Getting Email and Password from user and storing it in sqlite3 db which if not existent will be created by default and then after submitting the form, the user is redirected.

Condition to check if the user is Registered and the credentials match to sign in is yet to be created.